statistics: ADD kda, win rate service
ADD getKDAFromNormalMatch method to StaNormalMatchService
ADD getWinRateFromNormalMatch method to StaNormalMatchService
ADD StaNormalMatchServiceTest

TODO - extending kda, win rate service to remain repsository

+++ katarina-batch/batch-statistics/src/main/kotlin/com/hubtwork/katarina/statistics/statisticsapi/domain/StaARAM.kt
@@ -7,9 +7,9 @@ import javax.persistence.*
 class StaARAM(
     accountId: String,
     championId: Int,
-    kill: Int,
-    death: Int,
-    assist: Int,
+    kill: Float,
+    death: Float,
+    assist: Float,
     gameAllCount: Int,
     gameWinCount: Int,
     season: Int?,
@@ -19,7 +19,7 @@ class StaARAM(
     @Id
     @Column(name = "ARAM_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var ARAMId : Int = 0
+    var ARAMId : Long = 0
 
     @Column(name = "account_id")
     var accountId = accountId
